### 2024-11-01 #7419 @editorialbot
[PRE REVIEW]: votess: A multi-target, GPU-capable, parallel Voronoi tessellator
**Submitting author:** @samridh-dev (<a href="http://orcid.org/0009-0008-8620-3648">Samridh Dev Singh</a>)
**Repository:** https://github.com/samridh-dev/votess.git
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
