### 2024-11-01 #7419 @danielskatz
@samridh-dev - thanks for your submission. Our LOC counter shows 300k LOC of C++ in your repository. It seems likely that this is not all your code, but that you are including code from other packages. Can you help me understand what part of this is your code that we should be reviewing, and what is preexisting code that we would not review?
