### 2024-11-01 #7419 @danielskatz
Which part of the code (meaning which directories) should be counted and reviewed?
