### 2024-11-01 #7419 @samridh-dev
Only the src/ and include/ directories should be reviewed.
