+++ .github/workflows/joss.yml
@@ -16,7 +16,7 @@ jobs:
         with:
           journal: joss
           paper-path: joss/paper.md
-      - name: Upload
+      - name: Upload PDF
         uses: actions/upload-artifact@v4
         with:
           name: paper
@@ -25,4 +25,14 @@ jobs:
         uses: EndBug/add-and-commit@v9
         with:
           message: '(auto) Paper PDF Draft'
-          add: joss/*.pdf
+          add: joss/paper.pdf
+      - name: Upload TeX
+        uses: actions/upload-artifact@v4
+        with:
+          name: paper
+          path: joss/paper.tex
+      - name: Commit PDF to repository
+        uses: EndBug/add-and-commit@v9
+        with:
+          message: '(auto) Paper PDF Draft'
+          add: joss/paper.tex
