### 2024-11-15 #7419 @Kevin-Mattheus-Moerman
@BrunoLevy @lrineau @sloriot would you be interested in reviewing this submission for [JOSS](https://joss.theoj.org/) entitled: "votess: A multi-target, GPU-capable, parallel Voronoi tessellator"?.

[JOSS reviews](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) take place here on GitHub, and focus on [the software](https://github.com/samridh-dev/votess.git) as well [a short paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.07419/joss.07419/10.21105.joss.07419.pdf).

If you are interested please let me know here. Thanks!
