+++ joss/paper.bib
@@ -37,6 +37,8 @@
 @article{rycroft2009voro,
   title={VORO++: A three-dimensional Voronoi cell library in C++},
   author={Rycroft, Chris},
+  doi={10.2172/946741},
+  url={https://doi.org/10.2172/946741}
   year={2009}
 }
 
@@ -47,7 +49,7 @@ volume = {115},
 pages = {102995},
 year = {2023},
 issn = {0167-8191},
-doi = {https://doi.org/10.1016/j.parco.2023.102995},
+doi = {10.1016/j.parco.2023.102995},
 url = {https://www.sciencedirect.com/science/article/pii/S0167819123000017},
 author = {Guoqing Wu and Hongyun Tian and Guo Lu and Wei Wang},
 keywords = {Computational geometry, Voronoi tessellation, Parallel algorithm, Distributed-memory},
