+++ joss/paper.bib
@@ -38,7 +38,7 @@
   title={VORO++: A three-dimensional Voronoi cell library in C++},
   author={Rycroft, Chris},
   doi={10.2172/946741},
-  url={https://doi.org/10.2172/946741}
+  url={https://doi.org/10.2172/946741},
   year={2009}
 }
 
