+++ CONTRIBUTING.md
@@ -62,7 +62,7 @@ possible.
 - To see if other users have experienced (and potentially already solved) the
   same issue you are having, check if there is not already a bug report
   existing for your bug or error in the [bug
-  tracker](https://github.com/samridh-dev/votess.git/issues?q=label%3Abug).
+  tracker](https://github.com/samridh-dev/votess/issues?q=label%3Abug).
 - Also make sure to search the internet (including Stack Overflow) to see if
   users outside of the GitHub community have discussed the issue.
 - Collect information about the bug:
