### 2025-01-14 samridh-dev/votess#12 @samridh-dev
JOSS: visualizations
I have been doing a few attempts to creating visualizations for the 3D Voronoi Geometries produced by the program, and I have so far created two objects:
 - A single convex hull produced by random points lying on the surface of a sphere
 - A Tesellation of 1e5 randomly generated points
The images produced are below:
![random-pointset-1e5](https://github.com/user-attachments/assets/a2c7c344-8d0d-4461-96bb-26b5daa972bd)
![convexhull](https://github.com/user-attachments/assets/025adbd6-c559-41e4-ba2e-7873510ba277)
Let me know what I can do to make these visuals better.
