### 2025-01-15 #7419 @samridh-dev
Apologies for the delay. I have addressed the above points, and I created a visualization showing what the program can do, and I am working on adding them to the paper and the README.md file. Please review it and let me know if it is sufficient.
