### 2025-01-28 #7419 @Kevin-Mattheus-Moerman
@samridh-dev thanks for working on those points. I see you added a bar-chart. Is there any visualizations of the Voronoi cells themselves? I would recommend adding this.
