### 2025-01-29 #7419 @sloriot
I'm sorry I wasn't available for a full review but I can still give a few comments that you'll be free to take into account. It's kind of surprising to cite as reference  CGAL 4.12 while the latest release if 6.0.1. Also I don't see any benchmark with the parallel CPU implementation of CGAL, at least for computing the dual Delaunay triangulation. I can run the benchmark myself if that helps, provided the author indicates how to run votess for a set of random points.
