### 2025-02-01 #7419 @Kevin-Mattheus-Moerman
@pmocz would you be interested in reviewing this submission for JOSS? See also ☝️
