### 2025-02-01 #7419 @pmocz
> [@pmocz](https://github.com/pmocz) would you be interested in reviewing this submission for JOSS? See also ☝️

@Kevin-Mattheus-Moerman Happy to help out with this one!
