### 2025-02-01 #7741 @editorialbot
[REVIEW]: votess: A multi-target, GPU-capable, parallel Voronoi tessellator
**Submitting author:** @samridh-dev (<a href="http://orcid.org/0009-0008-8620-3648">Samridh Dev Singh</a>)
**Repository:** https://github.com/samridh-dev/votess.git
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
