### 2025-02-20 #7741 @BrunoLevy
Hi Kevin, I'm mostly done with my review (TL;DR the software is interesting, but I think the evaluation part needs more work).
