### 2025-02-22 samridh-dev/votess#15 @DanielVandH
[JOSS] Comments
Hi, this is part of my review for https://github.com/openjournals/joss-reviews/issues/7741.

Overall I find this software really interesting and think it is a nice contribution to this area. I just have some comments.

- Reproducibility: Is there some code somewhere that shows how the plot at the top of p. 2 was produced? What is the distribution of the points used for comparing the different algorithms? Since it's not specified I assume it's spatially uniform, but maybe it would be useful to also have a comparison for a dataset that has more spatial heterogenity. Does it matter?
- References: On Lines 28, references should be given for the Bowyer-Watson algorithms. On Line 33, the security radius condition should have a citation that would elaborate on this concept (I also think it should be italicised not in quotes).

- Other:
    - The figure in the Performance section isn't labelled anything (i.e. there is no "Figure 1").
    - Are there any plots somewhere? https://github.com/samridh-dev/votess/issues/12 seems decent. I think you should include some code in the README to show how a user can plot these objects, with a reference to it in the paper (whether to include the plot itself would be up to you if you think it improves the readibility, but definitely in the README it would be nice) e.g. saying that visualisation is simple with `pyvista` (or whatever) as demonstrated in the project's README.
    - The sentence on Line 40 seems a bit out of place. I would make that your conclusion (combined with Line 47) following the paragraph on Lines 41-46.
