### 2025-03-16 #7741 @Kevin-Mattheus-Moerman
@pmocz you have one box unticked on references. I checked them just now. They seem in order in terms of reference format etc. Do you feel the authors have cited sufficient relevant work?
