+++ joss/paper.bib
@@ -101,3 +101,38 @@ eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/nme.5987},
 abstract = {Summary This paper presents a new scalable parallelization scheme to generate the 3D Delaunay triangulation of a given set of points. Our first contribution is an efficient serial implementation of the incremental Delaunay insertion algorithm. A simple dedicated data structure, an efficient sorting of the points, and the optimization of the insertion algorithm have permitted to accelerate reference implementations by a factor three. Our second contribution is a multithreaded version of the Delaunay kernel that is able to concurrently insert vertices. Moore curve coordinates are used to partition the point set, avoiding heavy synchronization overheads. Conflicts are managed by modifying the partitions with a simple rescaling of the space-filling curve. The performances of our implementation have been measured on three different processors: an Intel core-i7, an Intel Xeon Phi, and an AMD EPYC, on which we have been able to compute three billion tetrahedra in 53 seconds. This corresponds to a generation rate of over 55 million tetrahedra per second. We finally show how this very efficient parallel Delaunay triangulation can be integrated in a Delaunay refinement mesh generator, which takes as input the triangulated surface boundary of the volume to mesh.},
 year = {2019}
 }
+
+@article{boyer1,
+    author = {Bowyer, A.},
+    title = {Computing Dirichlet tessellations*},
+    journal = {The Computer Journal},
+    volume = {24},
+    number = {2},
+    pages = {162-166},
+    year = {1981},
+    month = {01},
+    abstract = {An efficient algorithm is proposed for computing the Dirichlet tessellation and Delaunay triangulation in a k dimensional Euclidean space (k≥2). The algorithm is designed in a way that should allow it to be extended to some of the simpler non-Euclidean metric spaces as well. The algorithm has been implemented in ISO FORTRAN by the author and execution time and stereoscopic pictures of the tessellation and triangulation are presented at the end of this paper.},
+    issn = {0010-4620},
+    doi = {10.1093/comjnl/24.2.162},
+    url = {https://doi.org/10.1093/comjnl/24.2.162},
+    eprint = {https://academic.oup.com/comjnl/article-pdf/24/2/162/967239/240162.pdf},
+}
+
+@article{watson1,
+    author = {Watson, D. F.},
+    title = {Computing the n-dimensional Delaunay tessellation with application to Voronoi polytopes*},
+    journal = {The Computer Journal},
+    volume = {24},
+    number = {2},
+    pages = {167-172},
+    year = {1981},
+    month = {01},
+    abstract = {The Delaunay tessellation in n-dimensional space is a space-filling aggregate of n-simplices. These n-simplices are the dual forms of the vertices in the commonly used Voronoi tessellation. Several efforts have been made to simulate the 2-dimensional Voronoi tessellation on the computer. Additional problems occur for the 3 and higher dimensional implementations but some of these can be avoided by alternatively computing the dual Delaunay tessellation. An algorithm that finds the topological relationships in these tessellations is given.},
+    issn = {0010-4620},
+    doi = {10.1093/comjnl/24.2.167},
+    url = {https://doi.org/10.1093/comjnl/24.2.167},
+    eprint = {https://academic.oup.com/comjnl/article-pdf/24/2/167/967258/240167.pdf},
+}
+
+
+
