+++ joss/paper.bib
@@ -134,5 +134,35 @@ year = {2019}
     eprint = {https://academic.oup.com/comjnl/article-pdf/24/2/167/967258/240167.pdf},
 }
 
+@article{opticaltransport,
+    author = {Levy, Bruno and Mohayaee, Roya and von Hausegger, Sebastian},
+    title = {A fast semidiscrete optimal transport algorithm for a unique reconstruction of the early Universe},
+    journal = {Monthly Notices of the Royal Astronomical Society},
+    volume = {506},
+    number = {1},
+    pages = {1165-1185},
+    year = {2021},
+    month = {06},
+    abstract = {We leverage powerful mathematical tools stemming from optimal transport theory and transform them into an efficient algorithm to reconstruct the fluctuations of the primordial density field, built on solving the Monge-Ampère-Kantorovich equation. Our algorithm computes the optimal transport between an initial uniform continuous density field, partitioned into Laguerre cells, and a final input set of discrete point masses, linking the early to the late Universe. While existing early universe reconstruction algorithms based on fully discrete combinatorial methods are limited to a few hundred thousand points, our algorithm scales up well beyond this limit, since it takes the form of a well-posed smooth convex optimization problem, solved using a Newton method. We run our algorithm on cosmological N-body simulations, from the AbacusCosmos suite, and reconstruct the initial positions of \$\\mathcal \{O\}(10^7)\$ particles within a few hours with an off-the-shelf personal computer. We show that our method allows a unique, fast, and precise recovery of subtle features of the initial power spectrum, such as the baryonic acoustic oscillations.},
+    issn = {0035-8711},
+    doi = {10.1093/mnras/stab1676},
+    url = {https://doi.org/10.1093/mnras/stab1676},
+    eprint = {https://academic.oup.com/mnras/article-pdf/506/1/1165/38933704/stab1676.pdf},
+}
 
+@article{weygaert,
+    author = {Levy, Bruno and Mohayaee, Roya and von Hausegger, Sebastian},
+    title = {A fast semidiscrete optimal transport algorithm for a unique reconstruction of the early Universe},
+    journal = {Monthly Notices of the Royal Astronomical Society},
+    volume = {506},
+    number = {1},
+    pages = {1165-1185},
+    year = {2021},
+    month = {06},
+    abstract = {We leverage powerful mathematical tools stemming from optimal transport theory and transform them into an efficient algorithm to reconstruct the fluctuations of the primordial density field, built on solving the Monge-Ampère-Kantorovich equation. Our algorithm computes the optimal transport between an initial uniform continuous density field, partitioned into Laguerre cells, and a final input set of discrete point masses, linking the early to the late Universe. While existing early universe reconstruction algorithms based on fully discrete combinatorial methods are limited to a few hundred thousand points, our algorithm scales up well beyond this limit, since it takes the form of a well-posed smooth convex optimization problem, solved using a Newton method. We run our algorithm on cosmological N-body simulations, from the AbacusCosmos suite, and reconstruct the initial positions of \$\\mathcal \{O\}(10^7)\$ particles within a few hours with an off-the-shelf personal computer. We show that our method allows a unique, fast, and precise recovery of subtle features of the initial power spectrum, such as the baryonic acoustic oscillations.},
+    issn = {0035-8711},
+    doi = {10.1093/mnras/stab1676},
+    url = {https://doi.org/10.1093/mnras/stab1676},
+    eprint = {https://academic.oup.com/mnras/article-pdf/506/1/1165/38933704/stab1676.pdf},
+}
 
