### 2025-04-14 samridh-dev/votess#12 @cbyrohl
This looks great. Would it be possible to put them side-by-side (cube left, single hull right; ~70%/30% width respectively) as a zoom-in (some similar to [here](https://matplotlib.org/stable/gallery/subplots_axes_and_figures/zoom_inset_axes.html)) on a single cell near the surface of the cube?

Some minor suggestions for finalization, primarily if creating said zoom-in plot:

- cube plot:
  - reduce seed points to a lower number (maybe between 1e3-1e4?) so that individual cells remain visible at reduced image width in README/paper.
  - in the plot description briefly mention the "removal" of cells (and how this is done) near the viewer-facing three cube faces.
- convex hull: make hull color consistent to color in cube

Providing a link to the render script in the plot description within the README.md sounds would be great.

What do you think?
