### 2025-05-27 samridh-dev/votess#12 @samridh-dev
Hi @cbyrohl. Sorry for the delay. I managed to move the visualization scripts to a separate repository: https://github.com/samridh-dev/votess-visualization.git. As for the zoom in, I was able to create it but I think it would be too difficult for me to do a zoom in for a single cell (I could work on multiple though).
