### 2025-05-27 samridh-dev/votess#12 @samridh-dev
![Image](https://github.com/user-attachments/assets/4e8f4919-661d-47af-81f3-3bd762c13f11)
This is what the zoom in looks like now.
