### 2025-06-09 #7741 @samridh-dev
Hi @Kevin-Mattheus-Moerman, yes as for the update: We have created a new repository for benchmarking: https://github.com/samridh-dev/votess-benchmark, and we were spending time including CGAL into our benchmarks. We need to run it on a HPC cluster, and we will then add it into the draft. Changes in draft for the new required citations has also been added:  https://github.com/samridh-dev/votess/tree/main/joss. Lastly, some input was given for the visualizations here: https://github.com/samridh-dev/votess/issues/12, where I am aiming to address the suggestions sometime in the next few days.
