### 2025-07-15 a065e6f
updated paper.md
