### 2025-08-05 #7741 @DanielVandH
Is there a summary of these changes/responses in reply to the reviewer questions somewhere @samridh-dev?
