### 2025-08-15 samridh-dev/votess#15 @samridh-dev
- I have the benchmarks along with the plots over at this repository: https://github.com/samridh-dev/votess-benchmark.git. We will assume spatially uniform datasets, as heterogenous datasets would be outside the scope of the project.
- I have reworked the citations throughout the paper, so feel free to double check if I missed anything.
