### 2025-08-15 samridh-dev/votess#15 @DanielVandH
Thanks for the reply.

- On Line 41, do you mean to say "lies outside **the scope** of both..."?
- You should refer to the existences of the benchmarks repository for the source code for the plot, either in your repository's README or in the paper.
- My second-last point about plotting doesn't seem to be addressed.
- For multiple consecutive references, instead of doing e.g. ```[@ref1] [@ref2]```, you should do ```[@ref1; @ref2]``` to format them better.
- There is still no label for the figure in the performance section.
