### 2025-08-15 #7741 @BrunoLevy
Hello,
- Since the contribution is an implementation of Ray et.al's algorithm, it is absolutely necessary to compare to Ray et.al's implementation that is available [here](https://github.com/basselin7u/VoroGPU) as already indicated in my review.
- I think it is problematic that the GPU implementation is significantly slower than one of the reference multicore CPU implementations (CGAL) (BTW, comparison with the two other state of the art and readily available implementations of Delaunay listed in my review are still missing).
I think it would be OK if it was *slightly* slower (then it can be interesting to have a Delaunay running on the GPU if it is part of a larger method with other parts running on the GPU), but here it is almost one order of magnitude slower, so I think it is a serious obstacle to the usefulness of the proposed implementation in its current state.
