### 2025-10-30 #7741 @Kevin-Mattheus-Moerman
@samridh-dev please update us by next Monday November 3rd. If we do not hear from you by then we'll assume you no longer wish to pursue publishing this work with JOSS and we'll need to reject the paper.
